refactor: split implementation into files
